Consolidate import of shared interface
