+++ dashboard/api/src/functions/getReports.ts
@@ -1,7 +1,8 @@
 import { app, HttpRequest, HttpResponseInit, InvocationContext } from "@azure/functions";
-import { enumerateBlobs, getBlobContent, BlobTree, BlobTreeNode } from "../blobEnumerator";
+import { enumerateBlobs, getBlobContent, } from "../blobEnumerator";
 import { logRequestIdentity } from "../requestIdentity";
 import { SKILL_REPORT_PATTERN } from "../skillReport";
+import type { BlobTree, BlobTreeNode } from "../../../shared/blobTree";
 
 /**
  * Recursively collect all blob paths matching the SKILL-REPORT pattern from a tree node.
