+++ dashboard/sync/src/functions/syncMsbenchEvalMetrics.ts
@@ -1,7 +1,8 @@
 import { app, HttpRequest, HttpResponseInit, InvocationContext, Timer } from "@azure/functions";
 import { TableClient } from "@azure/data-tables";
 import { AzureCliCredential, ManagedIdentityCredential } from "@azure/identity";
-import { listMsbenchDates, enumerateMsbenchBlobs, getMsbenchBlobContent, BlobTreeNode } from "../msbenchBlobEnumerator";
+import { listMsbenchDates, enumerateMsbenchBlobs, getMsbenchBlobContent } from "../msbenchBlobEnumerator";
+import type { BlobTreeNode } from "../../../shared/blobTree";
 
 const MSBENCH_STORAGE_ACCOUNT = process.env.MSBENCH_STORAGE_ACCOUNT;
 const EVAL_TABLE_NAME = process.env.MSBENCH_EVAL_TABLE_NAME;
