+++ dashboard/sync/src/msbenchBlobEnumerator.ts
@@ -1,17 +1,6 @@
 import { BlobServiceClient } from "@azure/storage-blob";
 import { AzureCliCredential, ManagedIdentityCredential } from "@azure/identity";
-
-export interface BlobEntry {
-    name: string;
-    blobName: string;
-}
-
-export interface BlobTreeNode {
-    files: BlobEntry[];
-    children: Record<string, BlobTreeNode>;
-}
-
-export type BlobTree = Record<string, BlobTreeNode>;
+import type { BlobTree, BlobTreeNode } from "../../shared/blobTree";
 
 const MSBENCH_STORAGE_ACCOUNT = process.env.MSBENCH_STORAGE_ACCOUNT;
 const MSBENCH_REPORTS_CONTAINER_NAME = process.env.MSBENCH_REPORTS_CONTAINER;
